requirements.txt
MetaTrader5>=5.0.45
websocket-client>=1.6.0
aiohttp>=3.8.0
asyncio>=3.4.3
python-dotenv>=1.0.0
requests>=2.31.0
schedule>=1.2.0
colorlog>=6.7.0
pydantic>=2.0.0
cryptography>=41.0.0
prometheus-cliente=0.17.0
fastapie=0.87.0
uvicorn>=0.20.0
pandas>=1.5.0
numpy>=1.24.0
psutil>=5.9.0
